+++ packages/talker/lib/src/talker.dart
@@ -58,6 +58,7 @@ class Talker {
     _observer = observer ?? const _DefaultTalkerObserver();
     _errorHandler = errorHandler ?? TalkerErrorHandler(this.settings);
     _history = history ?? DefaultTalkerHistory(this.settings);
+    _historyFilter = _logger.filter;
   }
 
   void _initLogger(TalkerLogger? logger) {
@@ -86,6 +87,7 @@ class Talker {
   late TalkerFilter _filter;
   late TalkerObserver _observer;
   late TalkerHistory _history;
+  late LoggerFilter _historyFilter;
 
   // final _fileManager = FileManager();
 
@@ -131,6 +133,7 @@ class Talker {
     _logger = logger ?? _logger;
     _errorHandler = errorHandler ?? TalkerErrorHandler(this.settings);
     _history = DefaultTalkerHistory(this.settings, history: _history.history);
+    _historyFilter = _logger.filter;
   }
 
   final _talkerStreamController = StreamController<TalkerData>.broadcast();
@@ -368,13 +371,15 @@ class Talker {
     AnsiPen? pen,
   }) {
     final type = TalkerLogType.fromLogLevel(logLevel);
+    final penByLogKey = settings.getPenByLogKey(type.key);
+    final title = settings.getTitleByLogKey(type.key);
     final data = TalkerLog(
       key: type.key,
       message?.toString() ?? '',
-      title: settings.getTitleByLogKey(type.key),
+      title: title,
       exception: exception,
       stackTrace: stackTrace,
-      pen: pen ?? settings.getPenByLogKey(type.key),
+      pen: pen ?? penByLogKey,
       logLevel: logLevel,
     );
     _handleLogData(data);
@@ -394,6 +399,7 @@ class Talker {
       _logger.log(
         data.generateTextMessage(timeFormat: settings.timeFormat),
         level: data.logLevel ?? LogLevel.error,
+        pen: data.pen,
       );
     }
   }
@@ -402,23 +408,22 @@ class Talker {
     TalkerLog data, {
     LogLevel? logLevel,
   }) {
-    if (!settings.enabled) {
-      return;
-    }
+    /// If the Talker is disabled by settings
+    if (!settings.enabled) return;
 
+    /// If the log is not approved by the filter
     final isApproved = _isApprovedByFilter(data);
-    if (!isApproved) {
-      return;
-    }
+    if (!isApproved) return;
+
+    var pen = data.pen;
 
     final logTypeKey = data.key;
     if (logTypeKey != null) {
       data.title = settings.getTitleByLogKey(logTypeKey);
-      data.pen = settings.getPenByLogKey(
-        logTypeKey,
-        fallbackPen: data.pen,
-      );
+      pen = settings.getPenByLogKey(logTypeKey, fallbackPen: data.pen);
+      data.pen = pen;
     }
+
     _observer.onLog(data);
     _talkerStreamController.add(data);
     _handleForOutputs(data);
@@ -426,13 +431,17 @@ class Talker {
       _logger.log(
         data.generateTextMessage(timeFormat: settings.timeFormat),
         level: logLevel ?? data.logLevel,
-        pen: data.pen,
+        pen: pen,
       );
     }
   }
 
   void _handleForOutputs(TalkerData data) {
-    _history.write(data);
+    if (_historyFilter.shouldLog(
+        data.generateTextMessage(timeFormat: settings.timeFormat),
+        data.logLevel ?? LogLevel.debug)) {
+      _history.write(data);
+    }
   }
 
   bool _isApprovedByFilter(TalkerData data) {