+++ packages/talker_dio_logger/lib/dio_logs.dart
@@ -5,6 +5,7 @@ import 'package:talker/talker.dart';
 import 'package:talker_dio_logger/talker_dio_logger.dart';
 
 const _encoder = JsonEncoder.withIndent('  ');
+const _hiddenValue = '*****';
 
 class DioRequestLog extends TalkerLog {
   DioRequestLog(
@@ -23,20 +24,27 @@ class DioRequestLog extends TalkerLog {
   String get key => TalkerLogType.httpRequest.key;
 
   @override
+  LogLevel get logLevel => settings.logLevel;
+
+  @override
   String generateTextMessage({
     TimeFormat timeFormat = TimeFormat.timeAndSeconds,
   }) {
     var msg = '[$title] [${requestOptions.method}] $message';
 
     final data = requestOptions.data;
-    final headers = requestOptions.headers;
+    final headers = Map.from(requestOptions.headers);
 
     try {
       if (settings.printRequestData && data != null) {
         final prettyData = _encoder.convert(data);
         msg += '\nData: $prettyData';
       }
+
       if (settings.printRequestHeaders && headers.isNotEmpty) {
+        // HTTP headers are case-insensitive by standard
+        _replaceHiddenHeaders(headers);
+
         final prettyHeaders = _encoder.convert(headers);
         msg += '\nHeaders: $prettyHeaders';
       }
@@ -44,6 +52,22 @@ class DioRequestLog extends TalkerLog {
       // TODO: add handling can`t convert
     }
     return msg;
+  }
+
+  void _replaceHiddenHeaders(Map<dynamic, dynamic> headers) {
+    // HTTP headers are case-insensitive by standard
+    final lowerCaseHeaders = <String, String>{};
+    headers.forEach((key, value) {
+      lowerCaseHeaders[key.toLowerCase()] = key;
+    });
+
+    for (final hiddenHeader in settings.hiddenHeaders) {
+      final lowerCaseHiddenHeader = hiddenHeader.toLowerCase();
+      if (lowerCaseHeaders.containsKey(lowerCaseHiddenHeader)) {
+        final originalHeader = lowerCaseHeaders[lowerCaseHiddenHeader]!;
+        headers[originalHeader] = _hiddenValue;
+      }
+    }
   }
 }
 
@@ -64,6 +88,9 @@ class DioResponseLog extends TalkerLog {
   String get key => TalkerLogType.httpResponse.key;
 
   @override
+  LogLevel get logLevel => settings.logLevel;
+
+  @override
   String generateTextMessage({
     TimeFormat timeFormat = TimeFormat.timeAndSeconds,
   }) {
@@ -72,8 +99,17 @@ class DioResponseLog extends TalkerLog {
     final responseMessage = response.statusMessage;
     final data = response.data;
     final headers = response.headers.map;
+    final redirects = response.redirects;
 
     msg += '\nStatus: ${response.statusCode}';
+
+    if (settings.printResponseTime) {
+      final responseTime = _getResponseTime(response.requestOptions);
+
+      if (responseTime != null) {
+        msg += '\nTime: $responseTime ms';
+      }
+    }
 
     if (settings.printResponseMessage && responseMessage != null) {
       msg += '\nMessage: $responseMessage';
@@ -87,6 +123,13 @@ class DioResponseLog extends TalkerLog {
       if (settings.printResponseHeaders && headers.isNotEmpty) {
         final prettyHeaders = _encoder.convert(headers);
         msg += '\nHeaders: $prettyHeaders';
+      }
+
+      if (settings.printResponseRedirects && redirects.isNotEmpty) {
+        final prettyRedirects = redirects.map((redirect) {
+          return '[${redirect.statusCode} ${redirect.method} - ${redirect.location}]';
+        }).join('\n');
+        msg += '\nRedirects:\n$prettyRedirects';
       }
     } catch (_) {
       // TODO: add handling can`t convert
@@ -112,6 +155,9 @@ class DioErrorLog extends TalkerLog {
   String get key => TalkerLogType.httpError.key;
 
   @override
+  LogLevel get logLevel => LogLevel.error;
+
+  @override
   String generateTextMessage({
     TimeFormat timeFormat = TimeFormat.timeAndSeconds,
   }) {
@@ -124,6 +170,14 @@ class DioErrorLog extends TalkerLog {
 
     if (statusCode != null) {
       msg += '\nStatus: ${dioException.response?.statusCode}';
+    }
+
+    if (settings.printResponseTime) {
+      final responseTime = _getResponseTime(dioException.requestOptions);
+
+      if (responseTime != null) {
+        msg += '\nTime: $responseTime ms';
+      }
     }
 
     if (settings.printErrorMessage && responseMessage != null) {
@@ -140,4 +194,17 @@ class DioErrorLog extends TalkerLog {
     }
     return msg;
   }
+}
+
+///
+/// Get response time
+///
+int? _getResponseTime(RequestOptions options) {
+  final triggerTime = options.extra[TalkerDioLogger.kDioLogsTimeStampKey];
+
+  if (triggerTime is int) {
+    return DateTime.now().millisecondsSinceEpoch - triggerTime;
+  }
+
+  return null;
 }