+++ packages/talker_dio_logger/test/logs_test.dart
@@ -52,6 +52,71 @@ void main() {
           result, contains('Headers: {\n  "Authorization": "Bearer Token"\n}'));
     });
 
+    test('logLevel should return settings logLevel', () {
+      final requestOptions = RequestOptions(path: '/test', method: 'GET');
+      final settings = TalkerDioLoggerSettings(
+        logLevel: LogLevel.info,
+      );
+      final dioRequestLog = DioRequestLog(
+        'Test message',
+        requestOptions: requestOptions,
+        settings: settings,
+      );
+
+      expect(dioRequestLog.logLevel, equals(LogLevel.info));
+    });
+
+    test(
+        'generateTextMessage should include redirects if printResponseRedirects is true',
+        () {
+      final requestOptions = RequestOptions(
+        path: '/test',
+        method: 'GET',
+      );
+      final settings = TalkerDioLoggerSettings(printResponseRedirects: true);
+      final dioRequestLog = DioResponseLog(
+        'Test message',
+        response: Response(
+          requestOptions: requestOptions,
+          redirects: [
+            RedirectRecord(200, 'GET', Uri.parse('about:blank')),
+            RedirectRecord(200, 'POST', Uri.parse('about:blank')),
+          ],
+        ),
+        settings: settings,
+      );
+
+      final result = dioRequestLog.generateTextMessage();
+
+      expect(
+          result,
+          contains(
+              'Redirects:\n[200 GET - about:blank]\n[200 POST - about:blank]'));
+    });
+
+    test(
+        'generateTextMessage should not include redirects if printResponseRedirects is false',
+        () {
+      final requestOptions = RequestOptions(
+        path: '/test',
+        method: 'GET',
+      );
+      final settings = TalkerDioLoggerSettings(printResponseRedirects: false);
+      final dioRequestLog = DioResponseLog(
+        'Test message',
+        response: Response(
+          requestOptions: requestOptions,
+          redirects: [
+            RedirectRecord(200, 'GET', Uri.parse('about:blank')),
+          ],
+        ),
+        settings: settings,
+      );
+
+      final result = dioRequestLog.generateTextMessage();
+      expect(result.contains('Redirects:'), isFalse);
+    });
+
     // Add more tests for DioRequestLog as needed
   });
 
@@ -103,6 +168,80 @@ void main() {
     });
 
     test(
+        'generateTextMessage should include response time if printResponseTime is true',
+        () {
+      final response = Response(
+        requestOptions: RequestOptions(path: '/test', method: 'GET', extra: {
+          TalkerDioLogger.kDioLogsTimeStampKey:
+              DateTime.now().millisecondsSinceEpoch,
+        }),
+        statusCode: 200,
+        statusMessage: 'OK',
+      );
+      final settings = TalkerDioLoggerSettings(printResponseTime: true);
+      final dioResponseLog = DioResponseLog(
+        'Test message',
+        response: response,
+        settings: settings,
+      );
+
+      final result = dioResponseLog.generateTextMessage();
+
+      expect(result, matches(RegExp(r'Time: \d+ ms')));
+    });
+
+    test(
+        'generateTextMessage should not include response time if printResponseTime is false',
+        () {
+      final response = Response(
+        requestOptions: RequestOptions(path: '/test', method: 'GET', extra: {
+          TalkerDioLogger.kDioLogsTimeStampKey:
+              DateTime.now().millisecondsSinceEpoch,
+        }),
+        statusCode: 200,
+        statusMessage: 'OK',
+      );
+      final settings = TalkerDioLoggerSettings(printResponseTime: false);
+      final dioResponseLog = DioResponseLog(
+        'Test message',
+        response: response,
+        settings: settings,
+      );
+
+      final result = dioResponseLog.generateTextMessage();
+
+      expect(result, isNot(contains('Time:')));
+    });
+
+    test(
+        'generateTextMessage error should include include response time if printResponseTime is true',
+        () {
+      final response = Response(
+        requestOptions: RequestOptions(path: '/test', method: 'GET', extra: {
+          TalkerDioLogger.kDioLogsTimeStampKey:
+              DateTime.now().millisecondsSinceEpoch,
+        }),
+        statusCode: 404,
+      );
+      final dioException = DioException(
+        requestOptions: RequestOptions(path: '/test', method: 'GET', extra: {
+          TalkerDioLogger.kDioLogsTimeStampKey:
+              DateTime.now().millisecondsSinceEpoch,
+        }),
+        response: response,
+        message: 'Error message',
+      );
+
+      final settings = TalkerDioLoggerSettings(printResponseTime: true);
+      final dioErrorLog = DioErrorLog('Error title',
+          dioException: dioException, settings: settings);
+
+      final result = dioErrorLog.generateTextMessage();
+
+      expect(result, matches(RegExp(r'Time: \d+ ms')));
+    });
+
+    test(
         'generateTextMessage should include headers if printResponseHeaders is true',
         () {
       final response = Response(
@@ -125,6 +264,23 @@ void main() {
               '    "application/json"\n'
               '  ]\n'
               '}'));
+    });
+
+    test('logLevel should return settings logLevel', () {
+      final response = Response(
+        requestOptions: RequestOptions(path: '/test', method: 'GET'),
+        statusCode: 200,
+      );
+      final settings = TalkerDioLoggerSettings(
+        logLevel: LogLevel.warning,
+      );
+      final dioResponseLog = DioResponseLog(
+        'Test message',
+        response: response,
+        settings: settings,
+      );
+
+      expect(dioResponseLog.logLevel, equals(LogLevel.warning));
     });
 
     // Add more tests for DioResponseLog as needed
@@ -261,6 +417,23 @@ void main() {
             '  ]\n'
             '}',
           ));
+    });
+
+    test('logLevel should always return error level', () {
+      final dioException = DioException(
+        requestOptions: RequestOptions(path: '/test', method: 'GET'),
+        message: 'Error message',
+      );
+      final settings = TalkerDioLoggerSettings(
+        logLevel: LogLevel.info,
+      );
+      final dioErrorLog = DioErrorLog(
+        'Error title',
+        dioException: dioException,
+        settings: settings,
+      );
+
+      expect(dioErrorLog.logLevel, equals(LogLevel.error));
     });
   });
 }