+++ packages/talker_flutter/lib/src/controller/talker_view_controller.dart
@@ -1,14 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:talker_flutter/src/utils/download_logs/download_nonweb_logs.dart'
-    if (dart.library.html) 'package:talker_flutter/src/utils/download_logs/download_web_logs.dart';
+import 'package:talker_flutter/src/utils/download_logs/donwload_logs.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
 /// Controller to work with [TalkerScreen]
 class TalkerViewController extends ChangeNotifier {
+  TalkerViewController({bool expandedLogs = true, isLogOrderReversed = true}) {
+    _expandedLogs = expandedLogs;
+    _isLogOrderReversed = isLogOrderReversed;
+  }
+
   BaseTalkerFilter _filter = BaseTalkerFilter();
 
-  var _expandedLogs = true;
-  bool _isLogOrderReversed = true;
+  late bool _expandedLogs;
+  late bool _isLogOrderReversed;
 
   /// Filter for selecting specific logs and errors
   BaseTalkerFilter get filter => _filter;