+++ packages/talker_flutter/lib/src/ui/talker_actions/talker_actions_bottom_sheet.dart
@@ -26,17 +26,12 @@ class TalkerActionsBottomSheet extends StatelessWidget {
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            ...actions
-                .asMap()
-                .entries
-                .map(
-                  (e) => _ActionTile(
-                    talkerScreenTheme: talkerScreenTheme,
-                    action: e.value,
-                    showDivider: e.key != actions.length - 1,
-                  ),
-                )
-                .toList(),
+            for (int i = 0; i < actions.length; i++)
+              _ActionTile(
+                talkerScreenTheme: talkerScreenTheme,
+                action: actions[i],
+                showDivider: i != actions.length - 1,
+              )
           ],
         ),
       ),
@@ -75,7 +70,7 @@ class _ActionTile extends StatelessWidget {
         ),
         if (showDivider)
           Divider(
-            color: talkerScreenTheme.textColor.withOpacity(0.2),
+            color: talkerScreenTheme.textColor.withValues(alpha: 0.2),
             height: 1,
           ),
       ],