+++ packages/talker_flutter/lib/src/ui/talker_screen.dart
@@ -11,6 +11,8 @@ class TalkerScreen extends StatelessWidget {
     this.itemsBuilder,
     this.appBarLeading,
     this.customSettings,
+    this.isLogsExpanded = true,
+    this.isLogOrderReversed = true,
   }) : super(key: key);
 
   /// Talker implementation
@@ -32,16 +34,25 @@ class TalkerScreen extends StatelessWidget {
   /// Optional custom settings
   final ValueNotifier<List<CustomSettingsGroup>>? customSettings;
 
+  ///{@macro talker_flutter_is_log_exapanded}
+  final bool isLogsExpanded;
+
+  ///{@macro talker_flutter_is_log_order_reversed}
+  final bool isLogOrderReversed;
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: theme.backgroundColor,
       body: TalkerView(
         talker: talker,
+        itemsBuilder: itemsBuilder,
         theme: theme,
         appBarTitle: appBarTitle,
         appBarLeading: appBarLeading,
         customSettings: customSettings,
+        isLogsExpanded: isLogsExpanded,
+        isLogOrderReversed: isLogOrderReversed,
       ),
     );
   }