+++ packages/talker_flutter/lib/src/ui/talker_settings/widgets/talker_setting_card.dart
@@ -42,12 +42,11 @@ class TalkerSettingsCard extends StatelessWidget {
                 fontWeight: FontWeight.w500,
               ),
             ),
-            trailing: trailing ??
-                CupertinoSwitch(
-                  value: enabled,
-                  trackColor: canEdit ? Colors.red : Colors.grey,
-                  onChanged: canEdit ? onChanged : null,
-                ),
+            trailing: CupertinoSwitch(
+              value: enabled,
+              inactiveTrackColor: canEdit ? Colors.red : Colors.grey,
+              onChanged: canEdit ? onChanged : null,
+            ),
           ),
         ),
       ),