+++ packages/talker_flutter/lib/src/ui/theme/talker_screen_theme.dart
@@ -27,6 +27,7 @@ class TalkerScreenTheme {
   Map<String, Color> get logColors {
     if (_customColors != null) {
       final customMap = Map<String, Color>.from(_defaultColors);
+      //ignore: unnecessary_non_null_assertion
       customMap.addAll(_customColors!);
       return customMap;
     }