+++ packages/talker_flutter/lib/src/utils/download_logs/download_logs_native.dart
@@ -10,9 +10,5 @@ Future<void> downloadFile(String logs) async {
   final file =
       await File('$dirPath/talker_logs_$fmtDate.txt').create(recursive: true);
   await file.writeAsString(logs);
-  await Share.shareXFiles(
-    <XFile>[
-      XFile(file.path),
-    ],
-  );
+  await Share.shareXFiles(<XFile>[XFile(file.path)]);
 }