+++ packages/talker_logger/lib/src/logger_web.dart
@@ -1,4 +1,6 @@
-import 'dart:html';
+import 'dart:js_interop';
+
+import 'package:web/web.dart';
 
 void outputLog(String message) =>
-    message.split('\n').forEach(window.console.log);
+    message.split('\n').forEach((it) => console.log(it.toJS));